Removing models from repo. Moving tools used in each to concerto source folder.
