--- models/azotobacter_vinelandii_dj/README.md
@@ -1,4 +1,0 @@
-# Updates to iDT1278.xml
-
-change_model_compartment.py -  corrects compartments in the model.
-The model diff report can be viewed in fix_compartment_diff.html.
