--- models/azotobacter_vinelandii_dj/data/essentiality/knockouts.csv
@@ -1,3 +1,0 @@
-gene,essential,comment
-Avin_24430,no,test
-Avin_12480,no,test
