--- models/azotobacter_vinelandii_dj/data/growth/pm1_pm2_biolog.csv
@@ -1,68 +1,0 @@
-exchange,uptake,growth
-EX_arab__L_e,10,False
-EX_glcr_e,10,False
-EX_succ_e,10,True
-EX_gal_e,10,True
-EX_asp__L_e,10,True
-EX_ala__D_e,10,True
-EX_man_e,10,False
-EX_galt_e,10,False
-EX_ser__D_e,10,True
-EX_sbt__D_e,10,True
-EX_glyc_e,10,True
-EX_fuc__L_e,10,False
-EX_glcur_e,10,True
-EX_glcn_e,10,True
-EX_glyc3p_e,10,True
-EX_xyl__D_e,10,True
-EX_lac__L_e,10,True
-EX_for_e,10,True
-EX_mnl_e,10,True
-EX_g6p_e,10,True
-EX_galctn__D_e,10,True
-EX_rib__D_e,10,True
-EX_rmn_e,10,True
-EX_fru_e,10,True
-EX_malt_e,10,True
-EX_melib_e,10,True
-EX_thymd_e,10,False
-EX_asn__L_e,10,False
-EX_12ppd__S_e,10,True
-EX_akg_e,10,True
-EX_lcts_e,10,True
-EX_sucr_e,10,True
-EX_uri_e,10,False
-EX_gln__L_e,10,True
-EX_g1p_e,10,True
-EX_f6p_e,10,True
-EX_malttr_e,10,True
-EX_adn_e,10,False
-EX_cit_e,10,True
-EX_inost_e,10,True
-EX_fum_e,10,True
-EX_ppa_e,10,True
-EX_galct__D_e,10,True
-EX_glyclt_e,10,True
-EX_ins_e,10,True
-EX_ser__L_e,10,True
-EX_thr__L_e,10,True
-EX_ala__L_e,10,True
-EX_acac_e,10,True
-EX_mal__D_e,10,False
-EX_lyx__L_e,10,False
-EX_pyr_e,10,True
-EX_galctn__L_e,10,False
-EX_galur_e,10,False
-EX_peamn_e,10,False
-EX_acnam_e,10,True
-EX_all__D_e,10,False
-EX_4abut_e,10,False
-EX_but_e,10,True
-EX_dca_e,10,True
-EX_hxa_e,10,True
-EX_5dglcn_e,10,True
-EX_tartr__L_e,10,True
-EX_his__L_e,10,False
-EX_lys__L_e,10,False
-EX_orn_e,10,False
-EX_dha_e,10,False
