--- models/azotobacter_vinelandii_dj/data/growth/pm3_biolog.csv
@@ -1,31 +1,0 @@
-exchange,uptake,growth
-EX_nh4_e,10,True
-EX_no2_e,10,True
-EX_no3_e,10,True
-EX_urea_e,10,True
-EX_ala__L_e,10,False
-EX_asn__L_e,10,True
-EX_asp__L_e,10,False
-EX_gln__L_e,10,True
-EX_his__L_e,10,True
-EX_lys__L_e,10,False
-EX_ser__L_e,10,False
-EX_thr__L_e,10,False
-EX_ala__D_e,10,False
-EX_ser__D_e,10,False
-EX_orn_e,10,False
-EX_peamn_e,10,False
-EX_ade_e,10,False
-EX_adn_e,10,False
-EX_cytd_e,10,False
-EX_csn_e,10,False
-EX_gua_e,10,False
-EX_gsn_e,10,False
-EX_thym_e,10,False
-EX_thymd_e,10,False
-EX_ura_e,10,False
-EX_uri_e,10,False
-EX_ins_e,10,False
-EX_xan_e,10,False
-EX_xtsn_e,10,False
-EX_alltn_e,10,False
