--- models/azotobacter_vinelandii_dj/data/media/glucose.csv
@@ -1,8 +1,0 @@
-exchange,uptake,comment
-EX_co2_e,1000.0,
-EX_h2o_e,1000.0,
-EX_h_e,1000.0,
-EX_nh4_e,1000.0,
-EX_o2_e,1000.0,
-EX_pi_e,1000.0,
-EX_glc__D_e,10.0,
