--- models/azotobacter_vinelandii_dj/data/media/min_media_no_nitrogen.csv
@@ -1,17 +1,0 @@
-exchange,uptake,
-EX_pi_e,1000.0
-EX_fe3_e,1000.0
-EX_mn2_e,1000.0
-EX_fe2_e,1000.0
-EX_zn2_e,1000.0
-EX_mg2_e,1000.0
-EX_ca2_e,1000.0
-EX_ni2_e,1000.0
-EX_cu2_e,1000.0
-EX_cobalt2_e,1000.0
-EX_mobd_e,1000.0
-EX_so4_e,1000.0
-EX_k_e,1000.0
-EX_o2_e,1000.0
-EX_cl_e,1000.0
-EX_pyr_e,1000.0,
