--- models/synechococcus_elongatus_pcc_7942/data/media/syn_min_media.csv
@@ -1,15 +1,0 @@
-exchange,uptake
-EX_hco3_e,1.4551742831403973
-EX_mg2_e,0.0006043083928022225
-EX_nh4_e,0.0031
-EX_co2_e,0.7190666730577998
-EX_leu__L_e,0.028364741054516637
-EX_cobalt2_e,2.758158572551524e-07
-EX_no3_e,0.4029488349165458
-EX_fe3_e,8.634960628764796e-05
-EX_so4_e,0.008394278254311832
-EX_h_e,1.8823881901014012
-EX_photon650_e,0.0006043083928022225
-EX_photon670_e,11.697656196156876
-EX_photon690_e,18.200781855573425
-EX_pi_e,0.023000000000002113
